Status updates
- Updated to v0.3.0 of addon-framework
- addon-framwork has code to do status from manifestwork
- updated probers in volsync addon controller to pull out
  installedCSV from operator subscription
- Removed obsolete controller that was previously created
  to self-manage status
